requirements.txt
numpy>=1.11.0
numexpr>=2.4.6
pandas_datareader==0.7.0
tables
requests
seaborn
matplotlib
flask
pyfolio==0.8.0
git+https://github.com/quantopian/empyrical
git+https://github.com/czielinski/portfolioopt
XlsxWriter